add: change test structure and fix typo in red black tree
