--- lib/src/main/kotlin/bst/tests/traversals/InOrderTests.kt
@@ -1,3 +1,0 @@
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Assertions.assertEquals
