+++ lib/src/test/kotlin/trees/RegularTreeTests.kt
